chore: move function interfaces to policy-spi to avoid circular dependencies
