+++ core/policy/policy-evaluator/build.gradle.kts
@@ -18,7 +18,6 @@ plugins {
 }
 
 dependencies {
-    api(project(":spi:transfer-spi"))
 }
 publishing {
     publications {
