+++ core/transfer/build.gradle.kts
@@ -22,6 +22,7 @@ plugins {
 dependencies {
     api(project(":spi:policy-spi"))
     api(project(":spi:transfer-spi"))
+    api(project(":spi:contract-spi"))
     implementation(project(":common:state-machine-lib"))
     implementation(project(":common:util"))
     implementation("io.opentelemetry:opentelemetry-extension-annotations:${openTelemetryVersion}")
