+++ spi/policy-spi/build.gradle.kts
@@ -19,6 +19,7 @@ plugins {
 
 dependencies {
     api(project(":spi:core-spi"))
+    api(project(":spi:transfer-spi"))
 }
 
 publishing {
