+++ spi/policy-spi/src/main/java/org/eclipse/dataspaceconnector/spi/policy/evaluation/ResourceDefinitionRuleFunction.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.spi.policy;
+package org.eclipse.dataspaceconnector.spi.policy.evaluation;
 
 import org.eclipse.dataspaceconnector.policy.model.Rule;
 import org.eclipse.dataspaceconnector.spi.result.Result;
